fix(lint): remove unused import and rename file to snake_case
- Remove unused `* as m` import from LanguageSwitcher.svelte
- Rename sidebar-collapsed-item-variants.ts to sidebar_collapsed_item_variants.ts
