+++ src/lib/components/sidebar_collapsed_item_variants.ts
